Fix: use atomic to avoid data race in client (#254)
* fix: use atomic to avoid data race


* feature: temporary disable cdnsystem&manager test for CI


* feature: update .github/workflows/ci.yaml and codecov.yml


* feature: update ratelimiter test


* feature: update dynconfig test


* fix: ratelimiter and file utils test


* fix: ratelimiter and queue test


* fix: httputil test


* feat: remove httputils

+++ pkg/util/fileutils/test/file_utils_test.go
@@ -99,7 +99,7 @@ func (s *FileUtilsTestSuite) TestLink() {
 	st := statutils.GetSysStat(info)
 	s.Require().NotNil(st)
 
-	s.Require().Equal(uint16(2), st.Nlink)
+	s.Require().Equal(uint64(2), uint64(st.Nlink))
 
 	s.Require().True(fileutils.PathExist(s.testFile + ".link"))
